+++ R/pip_load_aux.R
@@ -13,7 +13,8 @@
 #' argument
 #' @param apply_label logical: if TRUE, predefined labels will apply to data
 #' loaded using `file_to_load` argument. Default TRUE. Tip: change to FALSE if
-#' the main strcuture of data has changed and labels have not been updated
+#' the main structure of data has changed and labels have not been updated
+#' @param verbose logical: whether to display message. Default is TRUE
 #' @inheritParams pip_find_cache
 #'
 #' @return
@@ -47,7 +48,9 @@ pip_load_aux <- function(measure     = NULL,
                                               measure, "/"),
                          version      = NULL,
                          file_to_load = NULL,
-                         apply_label  = TRUE) {
+                         apply_label  = TRUE,
+                         verbose      = getOption("pipload.verbose")
+                         ) {
 
 
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -215,7 +218,9 @@ pip_load_aux <- function(measure     = NULL,
   if (file.exists(file_to_load)) {
     df <- fst::read_fst(file_to_load,
                         as.data.table = TRUE)
-    cli::cli_alert_success("{load_msg}\n{.file {file_to_load}}")
+    if (verbose) {
+      cli::cli_alert_success("{load_msg}\n{.file {file_to_load}}")
+    }
 
   } else {
     msg <- paste("file `", measure, ".fst` does not exist.")
@@ -230,7 +235,9 @@ pip_load_aux <- function(measure     = NULL,
                              measure = measure)
 
   } else {
-    cli::cli_alert_info("Labels not applied to versioning data")
+    if (verbose) {
+      cli::cli_alert_info("Labels not applied to versioning data")
+    }
   }
   return(df)
 }