+++ R/pip_load_cache.R
@@ -29,7 +29,7 @@ pip_load_cache <- function(country          = NULL,
                            type             = c("dataframe", "list"),
                            root_dir         = Sys.getenv("PIP_ROOT_DIR"),
                            pipedir          = pip_create_globals(root_dir)$PIP_PIPE_DIR,
-                           verbose          = TRUE
+                           verbose      = getOption("pipload.verbose")
                            ) {
 
   #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~