+++ NEWS.md
@@ -1,4 +1,8 @@
 # pipload (development version)
+* Fix messaging in `pip_load_results()`
+
+# pipload 0.1.5
+* add `verbose` parameter
 
 # pipload 0.1.4
 * Improve messages when `root_dir` is not available